Better check for 61 dates

+++ biffy/__init__.py
@@ -16,7 +16,7 @@ def convert_date(date):
   from datetime import datetime, timedelta
   if long(date) == 0:
     return datetime(1900, 1, 1, 0, 0, 0) + timedelta(seconds=date * 24 * 60 * 60)
-  elif date >= 61.0:
+  elif long(date) >= 61:
     # According to Lotus 1-2-3, Feb 29th 1900 is a real thing, therefore we have to remove one day after that date
     return datetime(1899, 12, 31, 0, 0, 0) + timedelta(days=long(date) - 1, seconds=long((date % 1) * 24 * 60 * 60))
   else:
